All tests moved into the scalafmt-tests package
