+++ scalafmt-tests/src/test/scala/org/scalafmt/ScalafmtConfigTest.scala
