+++ scalafmt-tests/src/test/scala/org/scalafmt/ScalafmtRunnerTest.scala
