+++ scalafmt-tests/src/test/scala/org/scalafmt/ScalafmtTest.scala
