+++ scalafmt-tests/src/test/scala/org/scalafmt/cli/CliOptionsTest.scala
