+++ scalafmt-tests/src/test/scala/org/scalafmt/cli/CliTest.scala
