+++ scalafmt-tests/src/test/scala/org/scalafmt/cli/FakeGitOps.scala
