+++ scalafmt-tests/src/test/scala/org/scalafmt/cli/FileTestOps.scala
