+++ scalafmt-tests/src/test/scala/org/scalafmt/dynamic/PositionSyntax.scala
@@ -1,4 +1,4 @@
-package tests
+package org.scalafmt.dynamic
 
 import scala.meta._
