+++ scalafmt-tests/src/test/scala/org/scalafmt/dynamic/ScalafmtVersionSuite.scala
@@ -1,6 +1,5 @@
-package tests
+package org.scalafmt.dynamic
 
-import org.scalafmt.dynamic.ScalafmtVersion
 import org.scalafmt.dynamic.ScalafmtVersion.InvalidVersionException
 import org.scalatest.funsuite.AnyFunSuite
